+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/grocery/Item.kt
@@ -1,6 +1,6 @@
 package com.altmann.choresmanager.models.chores.grocery
 
-class Item(
+data class Item(
     val itemId: String,
     val name: String,
     var quantity: Int