+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/gym/Exercise.kt
@@ -1,6 +1,6 @@
 package com.altmann.choresmanager.models.chores.gym
 
-class Exercise(
+data class Exercise(
     val exerciseId: String,
     val name: String,
     val sets: Int?,